+++ druid/src/widget/mod.rs
@@ -53,7 +53,7 @@ mod view_switcher;
 mod widget;
 mod widget_ext;
 
-pub use self::image::{Image, ImageData};
+pub use self::image::Image;
 pub use align::Align;
 pub use button::Button;
 pub use checkbox::Checkbox;